move helpers from plugins/.../utils to scenario/utils
